Merged attribution changes from "MatlockMap" into OSMEmbedded.
The link in the "Leaflet" text was removed (because there's no "back"
button on a BB10 device, so no way of navigating back) and a "switch2osm"
note was added.  The text was massaged to fit along the bottom of the
screen on a BB Z10.

The alternative (of doing the attribution link "properly" by invoking the
browser) wasn't done because it would overcomplicate what is intended to
be the "simpest possible" example.